Changed the way error hadnling is done and added new Mutator routes
Changed the way the program handles errors, for a more optimal and standard way of handling errors. Added more meaningful text to each exception thrown.

Added new mutators routes, for updating the other parts of the task object.
Added a MutatorRequest class as a parameter object in the API route methods. This way there is a unified model that handles all mutations. To be implemented in other controllers later