added ResultUser model PasswordEncryptor and SignUp method in AuthContoller
Added SignUp method to the password controller.
Added a ResultUser model in the auth controller, that is being sent back as a response to the request maker. The model is secure and is making sure no sensitive data is returned back to the request maker.
Added a PasswordEncryptor class, part of the service layer, which is in charge of encrypting and decrypting passwords, when user creates or authenticates a profile. The passoword encryptor uses BCrypt algorithm for password hashing. The class has a method for hashing passwords and a method for verifing a hashed password.