Layed the scaffold of JWT and fixed a problem with PasswordEncryptor
Introduced JWT authentication in the project. The JWT will be the only way the application handles authentication and authorization. Registered JWT as a service in Program.cs. Added a JWTGenerator class that will be in charge of generating JWT.  JWT token is to be debbuged and implemented further later!
Moved the different Authentication objects from the AuthContoller and have them in the Models directory under Auth.

Fixed an issue with the PasswordEncryptor class, as the HashPassword method had a hardcoded value of password, that was glitching the normal flow of authentication.