Added docker, render.yaml and DelteTask controller
Added a docker file that will be utilised for the CD workflow of the application. Currently the CD workflow is not fully functional. The render.yaml file is needed for the deployment process in render.com. The application still needs a proper way of getting env variables.
 Implemented the missing DeleteTask controller, which is in charge of handling delete requests. for deleting tasks by taskKey.
Additionaly removed the "/" from the naming of the different routes in the API controllers